requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.1
mysql-connector-python==8.1.0
pytz==2024.1
scikit-learn==1.4.0
numpy>=1.21.0
pandas>=1.3.0
plotly>=5.3.0
streamlit>=1.10.0
python-dateutil>=2.8.2
pathlib>=1.0.1




